Fixed lab join and unjoin api

Approved-by: Hemarao Dugana

+++ app/Services/Manage/MemberManagementService.php
@@ -341,10 +341,6 @@ class MemberManagementService
                     $module_type = config('constants.member_management_component_type.lab_program');
                     $addedMemberResponse = __('responses.create_member_manger_success_lab_program');
                     break;
-                default:
-                    $module_type = null;
-                    $addedMemberResponse = null;
-                    break;
             }
             $auto_invite = config('constants.member_management_auto_invite.no');
             switch ($request->auto_invite) {
@@ -572,16 +568,18 @@ class MemberManagementService
         try {
             $module_type = self::getModuleType($component);
             $member = MemberManagement::whereIn('email', $request->email)->where(['module_id'=>$checkComponentBasedOnSlug->id, 'module_type'=>$module_type])->first();
-            $member_manger = MemberManagement::whereIn('email', $request->email)->where(['module_id'=>$checkComponentBasedOnSlug->id, 'module_type'=>$module_type])->delete();
-            if ($module_type == '1') {
-                $lab = LabService::getLabBasedOnId($member->module_id);
-                $request->organization_id = $lab->organization_id;
-                $request->privacy = $lab->privacy;
-                $request->title = $lab->title;
-                $request->category = $lab->category_id;
-                MixpanelHelper::mixpanel_tracking(config('mixpanel.leave_lab'), $request, auth()->user(), $request->ip());
-            }
-            if ($member_manger) {
+            if ($member) {
+                $member_manger = MemberManagement::whereIn('email', $request->email)->where(['module_id'=>$checkComponentBasedOnSlug->id, 'module_type'=>$module_type])->delete();
+                if ($module_type == '1') {
+                    $lab = LabService::getLabBasedOnId($member->module_id);
+                    $request = \Illuminate\Http\Request::capture();
+                    $request->organization_id = $lab->organization_id;
+                    $request->privacy = $lab->privacy;
+                    $request->title = $lab->title;
+                    $request->category = $lab->category_id;
+                    MixpanelHelper::mixpanel_tracking(config('mixpanel.leave_lab'), $request, auth()->user(), $request->ip());
+                }
+
                 return true;
             }
